fix(database): drop unsound era check in blockverify, fix 32-bit build
blockverify.checkEra used gledger.DetermineBlockType to independently
derive a decoded block's era from its header and reject a disagreement
with the caller's claimed type. DetermineBlockType classifies era from
the header's announced protocol-major version, which is a block
producer's hard-fork-readiness signal, not a record of which era the
bytes are actually encoded in: a producer announces the next era's
protocol major before that era's hard fork has triggered. This repo's
own immutable-chain testdata contains genuine Alonzo blocks carrying
protocol major 7 (Babbage's floor) that checkEra rejected as an era
mismatch -- a defect that recurs at every past and future hard-fork
boundary. Dropped the check; blockverify.Hash now verifies hash and
slot only, which already proves the returned bytes are the genuine,
requested content. Updated blockverify_test.go, benchmark_test.go
(loadMigrationFixtureBlock's scan-past-bad-blocks workaround is no
longer needed), storage_migration_test.go, and DATABASE.md to match.

Also fixed a 32-bit build break in ReassembleTxCbor: comparing
len(blockCbor) (int) directly against the untyped constant
math.MaxUint32 doesn't compile when int is 32 bits (GOARCH=386/arm).
Cast to uint64 before comparing.

+++ database/cbor_offset.go
@@ -328,7 +328,7 @@ func (t *TxCborParts) ReassembleTxCbor(blockCbor []byte) ([]byte, error) {
 	// caller passing an unexpectedly huge buffer fails loudly instead of
 	// silently wrapping into a bogus (and possibly smaller) blockLen that
 	// the bounds checks below would then validate against.
-	if len(blockCbor) > math.MaxUint32 {
+	if uint64(len(blockCbor)) > math.MaxUint32 {
 		return nil, fmt.Errorf(
 			"block CBOR length %d exceeds uint32 range",
 			len(blockCbor),
